@@ -2,6 +2,7 @@
 /**
  * @author Richard Weinhold
  */
+
 namespace ricwein\Crypto\Asymmetric;
 
 use ricwein\Crypto\Ciphertext;
@@ -9,6 +10,7 @@ use ricwein\Crypto\Exceptions\InvalidArgumentException;
 use ricwein\Crypto\Exceptions\UnexpectedValueException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
+use const SODIUM_CRYPTO_BOX_PUBLICKEYBYTES;
 
 /**
  * crypto-base for asymmetric keypair handling
@@ -19,7 +21,7 @@ abstract class CryptoBase
     /**
      * @var KeyPair|null
      */
-    protected $keypair = null;
+    protected ?KeyPair $keypair = null;
 
     /**
      * @param KeyPair|null $key
@@ -48,7 +50,7 @@ abstract class CryptoBase
     }
 
     /**
-     * @param  KeyPair $key
+     * @param KeyPair $key
      * @return self
      */
     public function loadKey(KeyPair $key): self
@@ -59,15 +61,18 @@ abstract class CryptoBase
 
     /**
      * load pub and secret keys as emphemeral keypair
-     * @param  string|KeyPair|null $pubKey
+     * @param string|KeyPair|null $pubKey
      * @return KeyPair
      * @throws UnexpectedValueException|InvalidArgumentException
+     * @throws \ricwein\Crypto\Exceptions\EncodingException
      */
     public function deriveKeyPair($pubKey = null): KeyPair
     {
         if ($this->keypair === null) {
             throw new UnexpectedValueException('asymmetric authentication crypto requires a keypair, but none given', 500);
-        } elseif ($this->keypair->getKey(KeyPair::PRIV_KEY) === null) {
+        }
+
+        if ($this->keypair->getKey(KeyPair::PRIV_KEY) === null) {
             throw new UnexpectedValueException('asymmetric authentication crypto requires a valid keypair, but is invalid', 500);
         }
 
@@ -90,30 +95,30 @@ abstract class CryptoBase
             }
             $pubKey = $pubKey->getKey(KeyPair::PUB_KEY);
         } elseif (!is_string($pubKey)) {
-            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is of type %s', \SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, is_object($pubKey) ? get_class($pubKey) : gettype($pubKey)), 400);
-        } elseif (mb_strlen($pubKey, '8bit') !== \SODIUM_CRYPTO_BOX_PUBLICKEYBYTES) {
-            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is %d bytes', \SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, mb_strlen($pubKey, '8bit')), 400);
+            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is of type %s', SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, is_object($pubKey) ? get_class($pubKey) : gettype($pubKey)), 400);
+        } elseif (mb_strlen($pubKey, '8bit') !== SODIUM_CRYPTO_BOX_PUBLICKEYBYTES) {
+            throw new InvalidArgumentException(sprintf('Encryption-public-key must be string of length %d bytes long, but is %d bytes', SODIUM_CRYPTO_BOX_PUBLICKEYBYTES, mb_strlen($pubKey, '8bit')), 400);
         }
 
         // create new Alice-Priv <=> Bob-Pub ephemeral KeyPair
         return new KeyPair([
-            KeyPair::PUB_KEY  => $pubKey,
+            KeyPair::PUB_KEY => $pubKey,
             KeyPair::PRIV_KEY => $privKey,
         ]);
     }
 
     /**
      * encrypt plaintext with libsodium authenticated asymmetric crypto
-     * @param  string              $plaintext
-     * @param  string|KeyPair|null $pubKey
+     * @param string $plaintext
+     * @param string|KeyPair|null $pubKey
      * @return Ciphertext
      */
     abstract public function encrypt(string $plaintext, $pubKey = null): Ciphertext;
 
     /**
      * decrypt ciphertext with libsodium authenticated asymmetric crypto
-     * @param  Ciphertext          $ciphertext
-     * @param  string|KeyPair|null $pubKey
+     * @param Ciphertext $ciphertext
+     * @param string|KeyPair|null $pubKey
      * @return string
      */
     abstract public function decrypt(Ciphertext $ciphertext, $pubKey = null): string;
@@ -121,18 +126,18 @@ abstract class CryptoBase
 
     /**
      * asymmetric File encryption using libsodium
-     * @param  File                $source
-     * @param  Storage|File|null   $destination
-     * @param  string|KeyPair|null $pubKey
+     * @param File $source
+     * @param Storage|File|null $destination
+     * @param string|KeyPair|null $pubKey
      * @return File
      */
     abstract public function encryptFile(File $source, $destination = null, $pubKey = null): File;
 
     /**
      * asymmetric File decryption using libsodium
-     * @param  File                $source
-     * @param  Storage|File|null   $destination
-     * @param  string|KeyPair|null $pubKey
+     * @param File $source
+     * @param Storage|File|null $destination
+     * @param string|KeyPair|null $pubKey
      * @return File
      */
     abstract public function decryptFile(File $source, $destination = null, $pubKey = null): File;
