@@ -2,18 +2,42 @@
 /**
  * @author Richard Weinhold
  */
+
 namespace ricwein\Crypto\Symmetric;
 
+use Exception;
+use ricwein\Crypto\Exceptions\EncodingException;
+use ricwein\Crypto\Exceptions\InvalidArgumentException;
+use ricwein\Crypto\Exceptions\UnexpectedValueException;
 use ricwein\Crypto\Helper;
 use ricwein\Crypto\Ciphertext;
 use ricwein\Crypto\Exceptions\KeyMismatchException;
 use ricwein\Crypto\Exceptions\MacMismatchException;
+use ricwein\FileSystem\Exceptions\ConstraintsException;
+use ricwein\FileSystem\Exceptions\Exception as FileSystemException;
+use ricwein\FileSystem\Exceptions\FileNotFoundException;
+use ricwein\FileSystem\Exceptions\RuntimeException;
+use ricwein\FileSystem\Exceptions\UnexpectedValueException as FileSystemUnexpectedValueException;
 use ricwein\FileSystem\File;
 use ricwein\FileSystem\Storage;
 use ricwein\FileSystem\Enum\Hash;
 use ricwein\FileSystem\Exceptions\UnsupportedException;
 use ricwein\FileSystem\Exceptions\AccessDeniedException;
 use ricwein\FileSystem\Storage\Extensions\Binary;
+use function array_shift;
+use function hash_equals;
+use function is_string;
+use function random_bytes;
+use function sodium_crypto_generichash_final;
+use function sodium_crypto_generichash_init;
+use function sodium_crypto_generichash_update;
+use function sodium_crypto_stream_xor;
+use function sodium_increment;
+use function sodium_memzero;
+use const SODIUM_CRYPTO_GENERICHASH_KEYBYTES;
+use const SODIUM_CRYPTO_SECRETBOX_NONCEBYTES;
+use const SODIUM_CRYPTO_STREAM_KEYBYTES;
+use const SODIUM_CRYPTO_STREAM_NONCEBYTES;
 
 /**
  * asymmetric Crypto using libsodium
@@ -24,70 +48,76 @@ class Crypto extends CryptoBase
     /**
      * @var int
      */
-    const FILE_BUFFER = 1048576; // == 1024^2 == 2^20
+    public const FILE_BUFFER = 1048576; // == 1024^2 == 2^20
 
     /**
-    * @inheritDoc
+     * @inheritDoc
      */
     public function encrypt(string $plaintext): Ciphertext
     {
         try {
 
             // generate nonce and HKDF salt
-            $nonce = \random_bytes(\SODIUM_CRYPTO_SECRETBOX_NONCEBYTES);
-            $salt  = \random_bytes(\SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
+            $nonce = random_bytes(SODIUM_CRYPTO_SECRETBOX_NONCEBYTES);
+            $salt = random_bytes(SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
 
             // split shared secret into authentication and encryption keys
-            list($encKey, $authKey) = $this->key->hkdfSplit($salt);
+            [$encKey, $authKey] = $this->key->hkdfSplit($salt);
 
             /**
              * encrypt actual plaintext (XSalsa20)
              * @see https://download.libsodium.org/doc/advanced/xsalsa20.html
              */
-            $encrypted = \sodium_crypto_stream_xor($plaintext, $nonce, $encKey);
-            \sodium_memzero($plaintext);
-            \sodium_memzero($encKey);
+            $encrypted = sodium_crypto_stream_xor($plaintext, $nonce, $encKey);
+            sodium_memzero($plaintext);
+            sodium_memzero($encKey);
 
             $ciphertext = new Ciphertext($encrypted, $salt, $nonce);
             $ciphertext->setAuthKey($authKey);
 
-            \sodium_memzero($encrypted);
-            \sodium_memzero($salt);
-            \sodium_memzero($nonce);
-            \sodium_memzero($authKey);
+            sodium_memzero($encrypted);
+            sodium_memzero($salt);
+            sodium_memzero($nonce);
+            sodium_memzero($authKey);
 
             // save used cipher algorithm for later
             $ciphertext->setCipher('xsalsa20');
 
             return $ciphertext;
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
             throw new KeyMismatchException('invalid security key', 500, $exception);
         }
     }
 
     /**
      * @inheritDoc
+     * @param Ciphertext $ciphertext
+     * @return string
+     * @throws MacMismatchException
+     * @throws EncodingException
+     * @throws InvalidArgumentException
+     * @throws UnexpectedValueException
      */
     public function decrypt(Ciphertext $ciphertext): string
     {
 
         // split shared secret into authentication and encryption keys
-        list($encKey, $authKey) = $this->key->hkdfSplit($ciphertext->getSalt());
+        [$encKey, $authKey] = $this->key->hkdfSplit($ciphertext->getSalt());
 
         // validate ciphertext mac
         if (!$ciphertext->setAuthKey($authKey)->isValidMac()) {
             throw new MacMismatchException('Invalid message authentication code', 400);
         }
 
-        \sodium_memzero($authKey);
+        sodium_memzero($authKey);
 
         // decrypt actual ciphertext
-        $plaintext = \sodium_crypto_stream_xor($ciphertext->getEncrypted(), $ciphertext->getNonce(), $encKey);
+        $plaintext = sodium_crypto_stream_xor($ciphertext->getEncrypted(), $ciphertext->getNonce(), $encKey);
 
-        \sodium_memzero($encKey);
+        sodium_memzero($encKey);
         unset($ciphertext);
 
-        if (!\is_string($plaintext)) {
+        if (!is_string($plaintext)) {
             throw new MacMismatchException('Invalid message authentication code', 400);
         }
 
@@ -95,10 +125,15 @@ class Crypto extends CryptoBase
     }
 
     /**
-     * @param  Storage           $source source-storage
-     * @param  File|Storage|null $destination given destination
+     * @param Storage $source source-storage
+     * @param File|Storage|null $destination given destination
      * @return File
-     * @throws UnsupportedException|AccessDeniedException
+     * @throws AccessDeniedException
+     * @throws UnsupportedException
+     * @throws ConstraintsException
+     * @throws FileSystemException
+     * @throws RuntimeException
+     * @throws FileSystemUnexpectedValueException
      */
     protected function prepareDestination(Storage $source, $destination = null): File
     {
@@ -114,9 +149,13 @@ class Crypto extends CryptoBase
 
         if (!$destination instanceof File) {
             throw new UnsupportedException(sprintf('unable to use file-base cryptography for the given destination-storage \'%s\'', get_class($destination)), 400);
-        } elseif ($destination->isFile() && !$destination->isWriteable()) {
+        }
+
+        if ($destination->isFile() && !$destination->isWriteable()) {
             throw new AccessDeniedException('unable to write to destination', 500);
-        } elseif (!$destination->isFile() && !$destination->touch(true)) {
+        }
+
+        if (!$destination->isFile() && !$destination->touch(true)) {
             throw new AccessDeniedException('unable to create destination file', 500);
         }
 
@@ -125,6 +164,19 @@ class Crypto extends CryptoBase
 
     /**
      * @inheritDoc
+     * @param File $source
+     * @param null $destination
+     * @return File
+     * @throws AccessDeniedException
+     * @throws ConstraintsException
+     * @throws FileNotFoundException
+     * @throws FileSystemException
+     * @throws FileSystemUnexpectedValueException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws RuntimeException
+     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function encryptFile(File $source, $destination = null): File
     {
@@ -143,16 +195,25 @@ class Crypto extends CryptoBase
     }
 
     /**
-     * @param  File $source
-     * @param  File $destination
+     * @param File $source
+     * @param File $destination
      * @return int bytes written
+     * @throws AccessDeniedException
+     * @throws ConstraintsException
+     * @throws FileNotFoundException
+     * @throws FileSystemUnexpectedValueException
+     * @throws InvalidArgumentException
      * @throws MacMismatchException
+     * @throws RuntimeException
+     * @throws UnexpectedValueException
+     * @throws UnsupportedException
+     * @throws Exception
      */
     protected function streamEncryptFile(File $source, File $destination): int
     {
         // generate (first) nonce and HKDF salt
-        $nonce = \random_bytes(\SODIUM_CRYPTO_STREAM_NONCEBYTES);
-        $salt = \random_bytes(\SODIUM_CRYPTO_STREAM_KEYBYTES);
+        $nonce = random_bytes(SODIUM_CRYPTO_STREAM_NONCEBYTES);
+        $salt = random_bytes(SODIUM_CRYPTO_STREAM_KEYBYTES);
 
         // split our key into authentication and encryption keys
         list($encKey, $authKey) = $this->key->hkdfSplit($salt);
@@ -165,21 +226,20 @@ class Crypto extends CryptoBase
         $destinationHandle = $destination->getHandle(Binary::MODE_WRITE);
 
         // write file-header
-        $destinationHandle->write($nonce, \SODIUM_CRYPTO_STREAM_NONCEBYTES);
-        $destinationHandle->write($salt, \SODIUM_CRYPTO_STREAM_KEYBYTES);
+        $destinationHandle->write($nonce, SODIUM_CRYPTO_STREAM_NONCEBYTES);
+        $destinationHandle->write($salt, SODIUM_CRYPTO_STREAM_KEYBYTES);
 
         // calculate initial mac
-        $mac = \sodium_crypto_generichash_init($authKey);
-        \sodium_crypto_generichash_update($mac, $nonce);
-        \sodium_crypto_generichash_update($mac, $salt);
+        $mac = sodium_crypto_generichash_init($authKey);
+        sodium_crypto_generichash_update($mac, $nonce);
+        sodium_crypto_generichash_update($mac, $salt);
 
-        \sodium_memzero($authKey);
-        \sodium_memzero($salt);
+        sodium_memzero($authKey);
+        sodium_memzero($salt);
 
         // fetch initial stats from source-file
         $size = $sourceHandle->getSize();
         $written = 0;
-        $readBytes = 0;
 
         // begin the streaming encryption
         while ($sourceHandle->remainingBytes() > 0) {
@@ -192,26 +252,26 @@ class Crypto extends CryptoBase
             }
 
             $read = $sourceHandle->read($readBytes);
-            $encrypted = \sodium_crypto_stream_xor($read, $nonce, $encKey);
+            $encrypted = sodium_crypto_stream_xor($read, $nonce, $encKey);
 
-            \sodium_crypto_generichash_update($mac, $encrypted);
+            sodium_crypto_generichash_update($mac, $encrypted);
             $written += $destinationHandle->write($encrypted);
 
-            \sodium_increment($nonce);
+            sodium_increment($nonce);
         }
 
-        \sodium_memzero($encKey);
-        \sodium_memzero($nonce);
+        sodium_memzero($encKey);
+        sodium_memzero($nonce);
 
         // Check that our input file was not modified before we MAC it
-        if (!\hash_equals($source->getHash(Hash::CONTENT, 'sha256'), $initHash)) {
+        if (!hash_equals($source->getHash(Hash::CONTENT, 'sha256'), $initHash)) {
             throw new MacMismatchException('read-only file has been modified since it was opened for reading', 500);
         }
 
         // finish encryption
         $written += $destinationHandle->write(
-            \sodium_crypto_generichash_final($mac, \SODIUM_CRYPTO_GENERICHASH_KEYBYTES),
-            \SODIUM_CRYPTO_GENERICHASH_KEYBYTES
+            sodium_crypto_generichash_final($mac, SODIUM_CRYPTO_GENERICHASH_KEYBYTES),
+            SODIUM_CRYPTO_GENERICHASH_KEYBYTES
         );
 
         // free handles
@@ -223,6 +283,19 @@ class Crypto extends CryptoBase
 
     /**
      * @inheritDoc
+     * @param File $source
+     * @param null $destination
+     * @return File
+     * @throws AccessDeniedException
+     * @throws ConstraintsException
+     * @throws FileNotFoundException
+     * @throws FileSystemException
+     * @throws FileSystemUnexpectedValueException
+     * @throws InvalidArgumentException
+     * @throws MacMismatchException
+     * @throws RuntimeException
+     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     public function decryptFile(File $source, $destination = null): File
     {
@@ -241,10 +314,17 @@ class Crypto extends CryptoBase
     }
 
     /**
-     * @param  File $source
-     * @param  File $destination
+     * @param File $source
+     * @param File $destination
      * @return int bytes written
+     * @throws AccessDeniedException
+     * @throws ConstraintsException
+     * @throws FileNotFoundException
+     * @throws InvalidArgumentException
      * @throws MacMismatchException
+     * @throws RuntimeException
+     * @throws UnexpectedValueException
+     * @throws UnsupportedException
      */
     protected function streamDecryptFile(File $source, File $destination): int
     {
@@ -255,28 +335,28 @@ class Crypto extends CryptoBase
         $sourceHandle->seek();
 
         // read file-header from encrypted file
-        $nonce = $sourceHandle->read(\SODIUM_CRYPTO_STREAM_NONCEBYTES);
-        $salt = $sourceHandle->read(\SODIUM_CRYPTO_STREAM_KEYBYTES);
+        $nonce = $sourceHandle->read(SODIUM_CRYPTO_STREAM_NONCEBYTES);
+        $salt = $sourceHandle->read(SODIUM_CRYPTO_STREAM_KEYBYTES);
 
         // split our key into authentication and encryption keys
-        list($encKey, $authKey) = $this->key->hkdfSplit($salt);
+        [$encKey, $authKey] = $this->key->hkdfSplit($salt);
 
-        $mac = \sodium_crypto_generichash_init($authKey);
-        \sodium_crypto_generichash_update($mac, $nonce);
-        \sodium_crypto_generichash_update($mac, $salt);
+        $mac = sodium_crypto_generichash_init($authKey);
+        sodium_crypto_generichash_update($mac, $nonce);
+        sodium_crypto_generichash_update($mac, $salt);
 
         // verify stream-mac
-        $chunkMacs = self::verifyStreamMac($sourceHandle, $mac);
+        $chunkMacs = $this->verifyStreamMac($sourceHandle, $mac);
 
-        \sodium_memzero($authKey);
-        \sodium_memzero($salt);
+        sodium_memzero($authKey);
+        sodium_memzero($salt);
 
         // fetch initial stats from source-file
-        $cipherEnd = $sourceHandle->getSize() - \SODIUM_CRYPTO_GENERICHASH_KEYBYTES;
+        $cipherEnd = $sourceHandle->getSize() - SODIUM_CRYPTO_GENERICHASH_KEYBYTES;
         $received = 0;
 
         // decrypt stream
-        while ($sourceHandle->remainingBytes() > \SODIUM_CRYPTO_GENERICHASH_KEYBYTES) {
+        while ($sourceHandle->remainingBytes() > SODIUM_CRYPTO_GENERICHASH_KEYBYTES) {
 
             // prevent overflow
             if (($sourceHandle->getPos() + self::FILE_BUFFER) > $cipherEnd) {
@@ -287,9 +367,9 @@ class Crypto extends CryptoBase
 
             $read = $sourceHandle->read($readBytes);
 
-            \sodium_crypto_generichash_update($mac, $read);
+            sodium_crypto_generichash_update($mac, $read);
             $calcMAC = Helper::safeStrcpy($mac);
-            $calc = \sodium_crypto_generichash_final($calcMAC, \SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
+            $calc = sodium_crypto_generichash_final($calcMAC, SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
 
             // someone attempted to add a chunk at the end.
             if (empty($chunkMacs)) {
@@ -297,27 +377,27 @@ class Crypto extends CryptoBase
             }
 
             // this chunk was altered after the original MAC was verified
-            $chunkMAC = \array_shift($chunkMacs);
-            if (!\hash_equals($chunkMAC, $calc)) {
+            $chunkMAC = array_shift($chunkMacs);
+            if (!hash_equals($chunkMAC, $calc)) {
                 throw new MacMismatchException('Invalid message authentication code', 400);
             }
 
             // this is where the decryption actually occurs
-            $decrypted = \sodium_crypto_stream_xor($read, $nonce, $encKey);
+            $decrypted = sodium_crypto_stream_xor($read, $nonce, $encKey);
             $received += $destinationHandle->write($decrypted);
 
-            \sodium_memzero($decrypted);
-            \sodium_increment($nonce);
+            sodium_memzero($decrypted);
+            sodium_increment($nonce);
         }
 
-        \sodium_memzero($encKey);
-        \sodium_memzero($nonce);
+        sodium_memzero($encKey);
+        sodium_memzero($nonce);
 
         // free handles
         $sourceHandle = null;
         $destinationHandle = null;
 
-        \sodium_memzero($mac);
+        sodium_memzero($mac);
         unset($chunkMacs);
 
         return $received;
@@ -326,19 +406,20 @@ class Crypto extends CryptoBase
 
     /**
      * read chunk-macs from stream
-     * @param  Binary $sourceHandle
-     * @param  string $mac
+     * @param Binary $sourceHandle
+     * @param string $mac
      * @return array
      * @throws MacMismatchException
+     * @throws RuntimeException
      */
     final private function verifyStreamMac(Binary $sourceHandle, string $mac): array
     {
         $start = $sourceHandle->getPos();
 
         // fetch hmac
-        $cipherEnd = $sourceHandle->getSize() - \SODIUM_CRYPTO_GENERICHASH_KEYBYTES;
+        $cipherEnd = $sourceHandle->getSize() - SODIUM_CRYPTO_GENERICHASH_KEYBYTES;
         $sourceHandle->seek($cipherEnd);
-        $storedMac = $sourceHandle->read(\SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
+        $storedMac = $sourceHandle->read(SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
         $sourceHandle->seek($start);
 
         $chunkMACs = [];
@@ -351,29 +432,29 @@ class Crypto extends CryptoBase
             // ciphertext? If so, only return a portion of the file
             if (($sourceHandle->getPos() + self::FILE_BUFFER) >= $cipherEnd) {
                 $break = true;
-                $read  = $sourceHandle->read($cipherEnd - $sourceHandle->getPos());
+                $read = $sourceHandle->read($cipherEnd - $sourceHandle->getPos());
             } else {
                 $read = $sourceHandle->read(self::FILE_BUFFER);
             }
 
             // We're updating our HMAC and nothing else
-            \sodium_crypto_generichash_update($mac, $read);
+            sodium_crypto_generichash_update($mac, $read);
 
             // Copy the hash state then store the MAC of this chunk
             /** @var string $chunkMAC */
-            $chunkMAC    = Helper::safeStrcpy($mac);
-            $chunkMACs[] = \sodium_crypto_generichash_final($chunkMAC, \SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
+            $chunkMAC = Helper::safeStrcpy($mac);
+            $chunkMACs[] = sodium_crypto_generichash_final($chunkMAC, SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
         }
 
         /**
          * We should now have enough data to generate an identical MAC
          */
-        $finalHMAC = \sodium_crypto_generichash_final($mac, \SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
+        $finalHMAC = sodium_crypto_generichash_final($mac, SODIUM_CRYPTO_GENERICHASH_KEYBYTES);
 
         /**
          * Use hash_equals() to be timing-invariant
          */
-        if (!\hash_equals($finalHMAC, $storedMac)) {
+        if (!hash_equals($finalHMAC, $storedMac)) {
             throw new MacMismatchException('Invalid message authentication code', 400);
         }
 
